rename the file containing the utilities for resampling
